Evaluation of AngelCopilot GPT

+++ Portfolio_Management/lib/utils.py
@@ -16,17 +16,17 @@ SMALL_SIZE = 20
 MEDIUM_SIZE = 23
 BIGGER_SIZE = 25
 
-plt.rc('font', size=SMALL_SIZE)  # controls default text sizes
-plt.rc('axes', titlesize=SMALL_SIZE)  # fontsize of the axes title
-plt.rc('axes', labelsize=MEDIUM_SIZE)  # fontsize of the x and y labels
-plt.rc('xtick', labelsize=SMALL_SIZE)  # fontsize of the tick labels
-plt.rc('ytick', labelsize=SMALL_SIZE)  # fontsize of the tick labels
-plt.rc('legend', fontsize=SMALL_SIZE)  # legend fontsize
-plt.rc('figure', titlesize=BIGGER_SIZE)  # fontsize of the figure title
+plt.rc("font", size=SMALL_SIZE)  # controls default text sizes
+plt.rc("axes", titlesize=SMALL_SIZE)  # fontsize of the axes title
+plt.rc("axes", labelsize=MEDIUM_SIZE)  # fontsize of the x and y labels
+plt.rc("xtick", labelsize=SMALL_SIZE)  # fontsize of the tick labels
+plt.rc("ytick", labelsize=SMALL_SIZE)  # fontsize of the tick labels
+plt.rc("legend", fontsize=SMALL_SIZE)  # legend fontsize
+plt.rc("figure", titlesize=BIGGER_SIZE)  # fontsize of the figure title
 
-sns.set_style('whitegrid')
+sns.set_style("whitegrid")
 
-plt.style.use('fivethirtyeight')
+plt.style.use("fivethirtyeight")
 
 
 def apply_perc_change(portfolio: List[pd.DataFrame]) -> List[pd.DataFrame]:
@@ -51,18 +51,18 @@ def compute_change(dataframe: pd.DataFrame) -> pd.DataFrame:
     :param dataframe: pandas DataFrame, the dataframe with the Index data
     :return: dataframe: pandas Dataframe, dataframe with computed change
     """
-    if 'Price' in dataframe.columns:
-        dataframe = dataframe[['Date', 'Open', 'Price']][dataframe['Open'] is not None]
-        dataframe['Open'] = dataframe['Open'].astype(float)
-        dataframe['Price'] = dataframe['Price'].astype(float)
-        dataframe['Change'] = dataframe['Price'] - dataframe['Open']
-        dataframe['Date'] = pd.to_datetime(dataframe['Date'])
+    if "Price" in dataframe.columns:
+        dataframe = dataframe[["Date", "Open", "Price"]][dataframe["Open"] is not None]
+        dataframe["Open"] = dataframe["Open"].astype(float)
+        dataframe["Price"] = dataframe["Price"].astype(float)
+        dataframe["Change"] = dataframe["Price"] - dataframe["Open"]
+        dataframe["Date"] = pd.to_datetime(dataframe["Date"])
     else:
-        dataframe = dataframe[['Date', 'Open', 'Adj Close']][dataframe['Open'] != 'null']
-        dataframe['Open'] = dataframe['Open'].astype(float)
-        dataframe['Adj Close'] = dataframe['Adj Close'].astype(float)
-        dataframe['Change'] = dataframe['Adj Close'] - dataframe['Open']
-        dataframe['Date'] = pd.to_datetime(dataframe['Date'])
+        dataframe = dataframe[["Date", "Open", "Adj Close"]][dataframe["Open"] != "null"]
+        dataframe["Open"] = dataframe["Open"].astype(float)
+        dataframe["Adj Close"] = dataframe["Adj Close"].astype(float)
+        dataframe["Change"] = dataframe["Adj Close"] - dataframe["Open"]
+        dataframe["Date"] = pd.to_datetime(dataframe["Date"])
 
     return dataframe
 
@@ -74,10 +74,10 @@ def compute_perc_change(dataframe: pd.DataFrame) -> pd.DataFrame:
     :param dataframe: pandas DataFrame, the dataframe with the Index data
     :return: pandas Dataframe, dataframe with computed percentage change
     """
-    if 'Price' in dataframe.columns:
-        dataframe['perc_change'] = dataframe['Price'].pct_change()
+    if "Price" in dataframe.columns:
+        dataframe["perc_change"] = dataframe["Price"].pct_change()
     else:
-        dataframe['perc_change'] = dataframe['Adj Close'].pct_change()
+        dataframe["perc_change"] = dataframe["Adj Close"].pct_change()
 
     return dataframe
 
@@ -94,7 +94,7 @@ def construct_portfolio_dictionary(
     :param description: List, List of strings with the Index descriptions
     :return: dictionary
     """
-    portfolio_dict = {'names': names, 'frames': dataframes_list, 'weights': weights, 'description': description}
+    portfolio_dict = {"names": names, "frames": dataframes_list, "weights": weights, "description": description}
 
     return portfolio_dict
 
@@ -105,12 +105,12 @@ def compute_return_since_inception(dataframe: pd.DataFrame) -> float:
     :param dataframe: pandas Dataframe, Dataframe with the Index data
     :return: float, the return since inception
     """
-    if 'Price' in dataframe.columns:
-        p0 = float(dataframe[dataframe['Date'] == dataframe['Date'].min()]['Price'])
-        p1 = float(dataframe[dataframe['Date'] == dataframe['Date'].max()]['Price'])
+    if "Price" in dataframe.columns:
+        p0 = float(dataframe[dataframe["Date"] == dataframe["Date"].min()]["Price"])
+        p1 = float(dataframe[dataframe["Date"] == dataframe["Date"].max()]["Price"])
     else:
-        p0 = float(dataframe[dataframe['Date'] == dataframe['Date'].min()]['Adj Close'])
-        p1 = float(dataframe[dataframe['Date'] == dataframe['Date'].max()]['Adj Close'])
+        p0 = float(dataframe[dataframe["Date"] == dataframe["Date"].min()]["Adj Close"])
+        p1 = float(dataframe[dataframe["Date"] == dataframe["Date"].max()]["Adj Close"])
 
     return (p1 - p0) / p0
 
@@ -127,7 +127,7 @@ def print_return_since_inception(dataframe: pd.DataFrame, name: str) -> None:
 
     print(
         f'The returns since inception {str(dataframe["Date"].min().date())} for "{name}" is '
-        f'{round(returns_since_inception * 100, 2)}%'
+        f"{round(returns_since_inception * 100, 2)}%"
     )
 
 
@@ -140,11 +140,11 @@ def find_intersection(list_df: List[pd.DataFrame]) -> Tuple[pd.Series, str, str]
 
     # Filter only non-null values
     for idx in range(len(list_df)):
-        list_df[idx] = list_df[idx][list_df[idx]['perc_change'].notnull()]
+        list_df[idx] = list_df[idx][list_df[idx]["perc_change"].notnull()]
 
-    intersection_list = set(list_df[0]['Date']).intersection(set(list_df[1]['Date']))
+    intersection_list = set(list_df[0]["Date"]).intersection(set(list_df[1]["Date"]))
     for df in list_df[2:]:
-        intersection_list = intersection_list.intersection(set(df['Date']))
+        intersection_list = intersection_list.intersection(set(df["Date"]))
 
     common_dates = pd.Series(list(intersection_list))
 
@@ -164,7 +164,7 @@ def intersect_dataframes(list_df: List[pd.DataFrame]) -> List[pd.DataFrame]:
     intersection = list_df.copy()
     common_dates, _, _ = find_intersection(intersection)
     for idx in range(len(intersection)):
-        intersection[idx] = intersection[idx][intersection[idx]['Date'].isin(common_dates)]
+        intersection[idx] = intersection[idx][intersection[idx]["Date"].isin(common_dates)]
 
     return intersection
 
@@ -176,11 +176,11 @@ def compute_mean_daily_returns(portfolio_dictionary: dict) -> List[float]:
     :param portfolio_dictionary: dictionary, the dictionary with the Index data
     :return: List[float], list with the mean daily returns of each Index
     """
-    intersected = intersect_dataframes(portfolio_dictionary['frames'])
+    intersected = intersect_dataframes(portfolio_dictionary["frames"])
     # Compute mean daily return per ETF
     mean_daily_returns = []
     for df in intersected:
-        mean_daily_returns.append(df['perc_change'].mean())
+        mean_daily_returns.append(df["perc_change"].mean())
 
     return mean_daily_returns
 
@@ -197,10 +197,10 @@ def compute_annualized_returns_no_intersection(portfolio_dictionary: dict) -> fl
     :return: float, the annualized return of the portfolio
     """
     mean_daily_returns = []
-    for df in portfolio_dictionary['frames']:
-        mean_daily_returns.append(df['perc_change'].mean())
+    for df in portfolio_dictionary["frames"]:
+        mean_daily_returns.append(df["perc_change"].mean())
 
-    weights = np.array(portfolio_dictionary['weights'])
+    weights = np.array(portfolio_dictionary["weights"])
 
     portfolio_returns = np.sum(mean_daily_returns * weights) * 252
